Socket-Server/requirements.txt
websockets
openai
python-dotenv
uvloop
